--- a/package.json
+++ b/package.json
@@ -209,6 +209,7 @@
   },
   "devDependencies": {
     "@markedjs/html-differ": "^2.0.1",
+    "autodll-webpack-plugin": "^0.4.2",
     "babel-core": "^6.26.3",
     "babel-eslint": "^10.0.1",
     "babel-loader": "^7.1.5",