| Q                | A
| ---------------- | ---
| Bug fix?         | no
| New feature?     |no
| License          | MIT

### Description

optimizing webpack compire time from 60s to <30s

|          | current | eslint-loader opt | autodll-webpack |
| -------- | ------- | ----------------- | --------------- |
| main     | 27s,28s | 28s,18s           | 3s,3s           |
| renderer | 30s,31s | 21s,19s           | 21s,20s         |
| total    | 57s,59s | 49s,37s           | 24s,23s         |

**note**
- i have test it in windows within dev and prod env,should test on mac or other env
- next we can autodll the nodejs env modules
- next we can autodll the muya assets libs js file