@Jocs 2019-05-25
@rainfu This PR significantly saves all contributors time, thank you very much for your contribution, but when I pull your code locally, the runtime reports an error, as follows:

```js
Uncaught ReferenceError: vendor_9894a12e780d86e7f1ef is not defined
    at eval (eval at dll-reference vendor_9894a12e780d86e7f1ef (renderer.js:10731), <anonymous>:1:18)
    at Object.dll-reference vendor_9894a12e780d86e7f1ef (renderer.js:10731)
    at __webpack_require__ (renderer.js:768)
    at fn (renderer.js:131)
    at eval (webpack-internal:///./node_modules/webpack/buildin/module.js:1)
    at Object../node_modules/webpack/buildin/module.js (renderer.js:5945)
    at __webpack_require__ (renderer.js:768)
    at fn (renderer.js:131)
    at eval (webpack-internal:///./node_modules/webpack-hot-middleware/client.js?noInfo=true&reload=true:308)
    at Object../node_modules/webpack-hot-middleware/client.js?noInfo=true&reload=true (renderer.js:5901)
```