@fxha 2019-05-25
Thank you for your contribution @rainfu, I'm not a JS expert but are there some reasons to use `autodll-webpack-plugin` instead of `hard-source-webpack-plugin` that is also used by webpack 5 like described in the readme?